requirements.txt
Pillow==10.1.0
PyPDF2==3.0.1
pytube==15.0.0
